requirements.txt
feedparser==6.0.11
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
PyYAML==6.0.1
SQLAlchemy==2.0.21
loguru==0.7.2
python-dotenv==1.0.0
tenacity==8.2.2
markdownify==0.11.6
dateparser==1.2.0
tiktoken==0.5.1
snscrape==0.7.0.20240311
streamlit==1.27.1
apscheduler==3.10.4
openai==1.6.1
pytest==7.4.3
